Build in repeatable container environment
I did not succeed to install the appropriate development packages on
my Fedora system (no ncurses'w'?), and kept getting missing symbols
during compilation, so I decided to try a repeatable container build
environment.  Ubuntu seemed as good as any choice here.  Fortunately
the binary works after copying it out of the container, and I am not
sure if it is statically linked or if the dynamically linked libraries
work anyway.

Signed-off-by: Gunnar Andersson <gunnar_dev@novaspring.eu>